Made types match.
svn:revision:7200
cvs:account:mike
cvs:branch:trunk

+++ mged/chgmodel.c
@@ -881,7 +881,7 @@ char	**argv;
 	struct loopuse *lu;
 	struct edgeuse *eu;
 	struct vertex *v_new, *v;
-	int tw, tf, tp;
+	unsigned long tw, tf, tp;
 	double log10();
 
 	rt_log("fracture:");
